refactor: extract reducer tab lifecycle
